# EDIT - DB를 사용하도록 수정 함.
 ### 수정수항
- 기존의 file을 통해 관리하던 정보를 DB를 통하여 하도록 수정

 ### 추가 및 기타 수정
- mysql-lib 사용
- mysql-lib에 mysql_read_all 함수 추가 (table의 모든 정보를 읽음)
- file 위치 이동 및 분할
