+++ src/JoinMerger.php
@@ -1,6 +1,6 @@
 <?php
-require_once('MsgHandler.php');
-require_once('../utils/dbQuery.php');
+require_once('../utils/mysql-lib.php');
+require_once('../utils/MsgHandler.php');
 
 if(empty($_POST)){
     return;
@@ -15,41 +15,20 @@ if(!checkMergerJson($json_data)){
     return;
 }
 
-$merger_id = dbQuote($json_data['mID']);
-$merger_list = dbSelectMerger($merger_id);
-$se_list = dbSelectSE();
+$result_arr = mysql_open_($json_data);
 
-$check_query = "SELECT * FROM merger WHERE mID = '$merger_id'";
-
-$check_result = dbQuery($check_query);
-
-$block_height = dbQuote($json_data['hgt']);
-$block_id = dbQuote($json_data['bID']);
-$prev_block_id = dbQuote($json_data['prevbID']);
-$prev_block_hash = dbQuote($json_data['prevHash']);
-$time = dbQuote($json_data['time']);
-
-if(mysqli_num_rows($check_result) > 0){
-    $update_query = "UPDATE 'merger' SET 
-                     'hgt' = '$block_height',
-                     'bID' = '$block_id',
-                     'prevbID' = '$prev_block_id',
-                     'prevHash' = '$prev_block_hash',
-                     
-                     'time' = '$time'";
-
-    dbQuery($update_query);
-}
-else{
-    $ip = dbQuote($json_data['ip']);
-    $port = dbQuote($json_data['port']);
-    $mCert = dbQuote($json_data['mCert']);
-
-    $insert_query = "INSERT INTO merger (mID, ip, port, mCert, time, bID, prevbID, prevHash )
-                    VALUES ('{$merger_id}', '{$ip}', '{$port}', '{$mCert}', '{$time}', '{$block_height}', '{$block_id}', '{$prev_block_id}', '{$prev_block_hash}' )";
-    
-    dbQuery($insert_query);
-}
+$merger_list = mysql_read_all('merger');
+$se_list = mysql_read_all('se');
 
 jsonResponse($merger_list, $se_list);
+
+$check_exist = mysql_read_urecord_('merger', 'mID', $result_arr['mID']);
+if(!$check_exist){
+    mysql_insert_('merger', $result_arr);
+}
+else{
+    mysql_update_urecordm_('merger', $result_arr, 'mID', $result_arr['mID']);
+}
+
+mysql_close();
